ai-service/requirements.txt
fastapi
uvicorn
pydantic

sqlalchemy
psycopg2-binary

langchain
langgraph

sentence-transformers
faiss-cpu

scikit-learn
numpy

openai